Impementar auth.js
